+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Fixed
+- Restored "Send diagnostic data" toggle to the redesigned Settings screen
+
 ## [2.7.0] - 2026-04-01
 
 ### Added